fix bug with empty glyphs
I misinterpreted what the purpose of checking the start/end offset of a
glyph was. If the diff is 0 then the glyph is empty and you don't have
to read it, but I was looking for the next different glyph offset. This
is okay if this happens before the end, but if it happened for the very
lst glyph, it would cause us to try and read past the end of the glyph
table.

Now we check if the view is 0 length and return an empty glyph instead.